Pub/Sub Fix again (#11)
fix(react-native): On emit events to the proper subRef, topic does not matter in this context. Filtering on topic only will publish multiple events to subscribers when multiple subscribers are listening to a single topic.

fix(ios): NSNumber did not coerce to Int and was being fed a garbage value. This prevented unsubscribing from a topic altogether.